refactor(ts-client): split opt-in features into extensions/
Move feature modules out of actions/ so the actions/ directory stays
scoped to core viem client decorators (public, wallet, encryption),
and add extensions/ as the home for opt-in feature modules.

- actions/depositContract.ts -> extensions/depositContract.ts
- actions/src20/             -> extensions/src20/
- faucet.ts                  -> extensions/faucet.ts

Import paths in client.ts and index.ts updated accordingly. No public
API changes.
