+++ clients/ts/packages/seismic-viem/src/extensions/faucet.ts
