+++ clients/ts/packages/seismic-viem/src/extensions/src20/src20Actions.ts
@@ -1,8 +1,8 @@
 import type { Account, Chain, Hex, Transport } from 'viem'
 
-import type { WatchSRC20EventsParams } from '@sviem/actions/src20/types.ts'
-import { watchSRC20Events } from '@sviem/actions/src20/watchSRC20Events.ts'
-import { watchSRC20EventsWithKey } from '@sviem/actions/src20/watchSRC20EventsWithKey.ts'
+import type { WatchSRC20EventsParams } from '@sviem/extensions/src20/types.ts'
+import { watchSRC20Events } from '@sviem/extensions/src20/watchSRC20Events.ts'
+import { watchSRC20EventsWithKey } from '@sviem/extensions/src20/watchSRC20EventsWithKey.ts'
 import type {
   ShieldedPublicClient,
   ShieldedWalletClient,
