+++ clients/ts/packages/seismic-viem/src/extensions/src20/watchSRC20EventsWithKey.ts
@@ -1,9 +1,9 @@
 import type { Address, Hex } from 'viem'
 
 import { SRC20Abi } from '@sviem/abis/src20.ts'
-import { parseEncryptedData } from '@sviem/actions/src20/crypto.ts'
-import { computeKeyHash } from '@sviem/actions/src20/directory.ts'
-import type { WatchSRC20EventsParams } from '@sviem/actions/src20/types.ts'
+import { parseEncryptedData } from '@sviem/extensions/src20/crypto.ts'
+import { computeKeyHash } from '@sviem/extensions/src20/directory.ts'
+import type { WatchSRC20EventsParams } from '@sviem/extensions/src20/types.ts'
 import type { ShieldedPublicClient } from '@sviem/client.ts'
 import { AesGcmCrypto } from '@sviem/crypto/aes.ts'
