+++ clients/ts/packages/seismic-viem/src/index.ts
@@ -72,8 +72,8 @@ export type {
 
 export type { ShieldedContract } from '@sviem/contract/contract.ts'
 
-export type { CheckFaucetParams } from '@sviem/faucet.ts'
-export { checkFaucet } from '@sviem/faucet.ts'
+export type { CheckFaucetParams } from '@sviem/extensions/faucet.ts'
+export { checkFaucet } from '@sviem/extensions/faucet.ts'
 
 export type {
   GetTxExplorerUrlParams,
@@ -128,25 +128,25 @@ export {
 export type { Secp256K1SigParams } from '@sviem/precompiles/secp256k1.ts'
 
 export type { CallClient, Precompile } from '@sviem/precompiles/precompile.ts'
-export { DEPOSIT_CONTRACT_ADDRESS } from '@sviem/actions/depositContract.ts'
+export { DEPOSIT_CONTRACT_ADDRESS } from '@sviem/extensions/depositContract.ts'
 
 // SRC20 event watching
-export { watchSRC20Events } from '@sviem/actions/src20/watchSRC20Events.ts'
-export { watchSRC20EventsWithKey } from '@sviem/actions/src20/watchSRC20EventsWithKey.ts'
+export { watchSRC20Events } from '@sviem/extensions/src20/watchSRC20Events.ts'
+export { watchSRC20EventsWithKey } from '@sviem/extensions/src20/watchSRC20EventsWithKey.ts'
 export {
   src20PublicActions,
   src20WalletActions,
-} from '@sviem/actions/src20/src20Actions.ts'
+} from '@sviem/extensions/src20/src20Actions.ts'
 export type {
   SRC20PublicActions,
   SRC20WalletActions,
-} from '@sviem/actions/src20/src20Actions.ts'
+} from '@sviem/extensions/src20/src20Actions.ts'
 export type {
   DecryptedTransferLog,
   DecryptedApprovalLog,
   WatchSRC20EventsParams,
   WatchSRC20EventsWithKeyParams,
-} from '@sviem/actions/src20/types.ts'
+} from '@sviem/extensions/src20/types.ts'
 
 // Directory contract helpers
 export {
@@ -155,5 +155,5 @@ export {
   getKey,
   registerKey,
   computeKeyHash,
-} from '@sviem/actions/src20/directory.ts'
+} from '@sviem/extensions/src20/directory.ts'
 export { DIRECTORY_ADDRESS, DirectoryAbi } from '@sviem/abis/directory.ts'
